\name{identity2}
\alias{identity2}
\title{
The Identity Arrow
}
\description{
The identity arrow for the Continuation category for which holds: f \%.\% identity2 = f = identity2 \%.\% f
}
\usage{
identity2(x, ret)
}
\arguments{
  \item{x}{
The value on which the function operates
}
  \item{ret}{
The following computation
}
}
\value{
This function always returns the original arrow.
}
\author{
Matteo Provenzano\cr
\url{http://www.alephdue.com}
}
